wip: logging & icmp receiver
added more spdlog logging throughout the code.
lots of updating of comments - still a lot more work to do to bring them up to scratch, but the base is there.
switched ICMPPingEngine to a single receiver which uses a signal to pass on packets to engine instances.
added string operator to ICMPPacket