wip: general
fixed moc includes not being in the include path.
removed sharedLibraryPath variable which was not used in the main app.
added context() function to IContextManager which returns the context id for the string based context id.
switched from using QLineEdit to QTextEdit to make use of syntax highlighting.
added functions to utils to check if host name and interval strings are valid
NewTargetDialog and NewTargetRibbonGroup now validate fields