feat: use offscreen bitmaps for graph background
performance of graph drawing gradually got worse to the point of locking up the GUI, this was due to the repeated drawing of the gradient background.  The route analyser now draws the background to an offscreen pixmap which is drawn to the graph.  The pixmap is updated when the max latency changes or the size of the graph changes.  A garbage collection is run to free dead pixmaps.