fix(TrimmerWidget.cpp): use QRectF instead of QRect
Switched the drawing to use QRectF, this solves a slight issue which when moving the start gripper, the end gripper would sometimes "jiggle"  by +- one pixel due to a rounding error.  Switching to QRectF preserves the fractional parts and therefore is unaffected by rounding resulting in a rock solid stable end point.