wip: general work
IEditor now has activated() and deactivated() methods which the editor manager calls when editors are switched.
Added EditorManagerTabWidget which is a subclass of QTabWidget that paints the background with a message or instruction to the user.
Added the ability to set the ordering of ribbon pages using unit values.
Added the ability so programatically select the current page on the ribbon bar.
Fixed ICMPPingEngine which was free-ing the singleton of the receiver worker, now that has moved to the destruction of the factory.